aws_strands_vs_agents_demo/requirements.txt
strands-agents
strands-agents-tools
strands-agents[anthropic]
python-dotenv